+++ packages/camera/camera_android_camerax/CHANGELOG.md
@@ -3,3 +3,4 @@
 * Creates camera_android_camerax plugin for development.
 * Adds CameraInfo class and removes unnecessary code from plugin.
 * Adds CameraSelector class.
+* Adds ProcessCameraProvider class.